+++ include/simpledbus/base/Exceptions.h
@@ -42,6 +42,15 @@ class InterfaceNotFoundException : public BaseException {
     std::string _message;
 };
 
+class PathNotFoundException : public BaseException {
+  public:
+    PathNotFoundException(const std::string& path, const std::string& subpath);
+    const char* what() const noexcept override;
+
+  private:
+    std::string _message;
+};
+
 }  // namespace Exception
 
 }  // namespace SimpleDBus